pyproject.toml
[project]
name = "flowstate-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.3",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "django>=6.0",
    "django-cors-headers>=4.5",
    "django-split-settings>=1.3",
    "djangorestframework>=3.16",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0",
    "psycopg[binary]>=3.2",
    "pydantic>=2.11",
    "pydantic-settings>=2.6",
    "uvicorn[standard]>=0.31",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15",
    "pytest>=8.0",
    "pytest-django>=4.9",
    "ruff>=0.10",
]
